+++ examples/register_gamepads.rs
@@ -1,6 +1,6 @@
 //! Demonstrates how to register gamepads in local multiplayer fashion
 
-use bevy::{prelude::*, utils::HashMap};
+use bevy::{platform::collections::HashMap, prelude::*};
 use leafwing_input_manager::prelude::*;
 
 fn main() {
@@ -59,10 +59,12 @@ fn join(
                     .id();
 
                 // Insert the created player and its gamepad to the hashmap of joined players
-                // Since uniqueness was already checked above, we can insert here unchecked
-                joined_players
-                    .0
-                    .insert_unique_unchecked(gamepad_entity, player);
+                // SAFETY: Since uniqueness was already checked above, we can insert here unchecked
+                unsafe {
+                    joined_players
+                        .0
+                        .insert_unique_unchecked(gamepad_entity, player);
+                }
             }
         }
     }