+++ src/action_state/mod.rs
@@ -4,16 +4,16 @@ use crate::input_map::UpdatedValue;
 use crate::{action_diff::ActionDiff, input_map::UpdatedActions};
 use crate::{Actionlike, InputControlKind};
 
+use bevy::platform::{collections::HashMap, time::Instant};
 use bevy::prelude::Resource;
 use bevy::reflect::Reflect;
-#[cfg(feature = "timing")]
-use bevy::utils::Duration;
-use bevy::utils::{HashMap, Instant};
 use bevy::{ecs::component::Component, prelude::ReflectComponent};
 use bevy::{
     math::{Vec2, Vec3},
     prelude::ReflectResource,
 };
+#[cfg(feature = "timing")]
+use core::time::Duration;
 use serde::{Deserialize, Serialize};
 
 mod action_data;
@@ -45,7 +45,7 @@ pub use action_data::*;
 /// ```rust
 /// use bevy::reflect::Reflect;
 /// use leafwing_input_manager::prelude::*;
-/// use bevy::utils::Instant;
+/// use bevy::platform::time::Instant;
 ///
 /// #[derive(Actionlike, PartialEq, Eq, Hash, Clone, Copy, Debug, Reflect)]
 /// enum Action {
@@ -171,7 +171,7 @@ impl<A: Actionlike> ActionState<A> {
     /// use bevy::prelude::Reflect;
     /// use leafwing_input_manager::prelude::*;
     /// use leafwing_input_manager::buttonlike::ButtonState;
-    /// use bevy::utils::Instant;
+    /// use bevy::platform::time::Instant;
     ///
     /// #[derive(Actionlike, Clone, Copy, PartialEq, Eq, Hash, Debug, Reflect)]
     /// enum Action {