+++ src/user_input/keyboard.rs
@@ -53,7 +53,6 @@ impl UpdatableInput for KeyCode {
 #[serde_typetag]
 impl Buttonlike for KeyCode {
     /// Checks if the specified key is currently pressed down.
-    #[must_use]
     #[inline]
     fn pressed(&self, input_store: &CentralInputStore, _gamepad: Entity) -> bool {
         input_store.pressed(self)
@@ -72,6 +71,7 @@ impl Buttonlike for KeyCode {
             state: ButtonState::Pressed,
             repeat: false,
             window: Entity::PLACEHOLDER,
+            text: None,
         });
     }
 
@@ -88,6 +88,7 @@ impl Buttonlike for KeyCode {
             state: ButtonState::Released,
             repeat: false,
             window: Entity::PLACEHOLDER,
+            text: None,
         });
     }
 
@@ -177,7 +178,6 @@ impl UserInput for ModifierKey {
 #[serde_typetag]
 impl Buttonlike for ModifierKey {
     /// Checks if the specified modifier key is currently pressed down.
-    #[must_use]
     #[inline]
     fn pressed(&self, input_store: &CentralInputStore, _gamepad: Entity) -> bool {
         input_store.pressed(&self.left()) || input_store.pressed(&self.right())